chatbot-service/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
cohere==5.9.1
qdrant-client==1.11.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.1
pydantic-settings==2.5.2
sqlalchemy==2.0.35
requests==2.32.3
google-generativeai==0.8.3
